# backend/Document/Swagger.yaml
openapi: 3.0.4
info:
  title: National Football League Management API
  description: |-
    This is an API for managing a national football league, built on the OpenAPI 3.0 specification. It provides endpoints for managing team information, players, match schedules, results, and league standings.  
    Learn more about OpenAPI at [https://swagger.io](https://swagger.io). We follow a design-first approach, allowing you to contribute to improving the API, from its definition to its code.  
    Some useful links:  
    - [National Football League API Repository](https://github.com/truonghoangkhiem/QL_Giai_BD_QG_SE104_UIT.git)
  termsOfService: https://swagger.io/terms/
  contact:
    email: truongkhiemhoang@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
...
servers:
  - url: http://localhost:5000
...
paths:
...
  /api/regulations/:
    get:
      tags:
        - regulations
      summary: Get all regulations
      description: Retrieves a list of all regulations
      operationId: getRegulations
      responses:
        "200":
          description: Regulations fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegulationsResponse"
    post:
      tags:
        - regulations
      summary: Create a new regulation
      description: Creates a new regulation for a season
      operationId: createRegulation
      security:
        - bearerAuth: []
      requestBody:
        description: Information needed to create a regulation
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegulationRequest"
      responses:
        "201":
          description: Regulation created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid input, missing field, invalid rules, or regulation already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/regulations/{id}:
    get:
      tags:
        - regulations
      summary: Get a regulation by ID
      description: Retrieves a regulation by its ID
      operationId: getRegulationById
      parameters:
        - $ref: "#/components/parameters/RegulationId"
      responses:
        "200":
          description: Regulation found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegulationResponse"
        "400":
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Regulation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - regulations
      summary: Update a regulation
      description: Updates the rules of an existing regulation
      operationId: updateRegulation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RegulationId"
      requestBody:
        description: Information needed to update the regulation
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRegulationRequest"
      responses:
        "200":
          description: Regulation updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid ID format or invalid rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Regulation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - regulations
      summary: Delete a regulation
      description: Deletes an existing regulation by its ID
      operationId: deleteRegulation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RegulationId"
      responses:
        "200":
          description: Regulation deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Regulation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/regulations/{season_id}/{regulation_name}:
    get:
      tags:
        - regulations
      summary: Get regulation ID by season and name
      description: Retrieves the ID of a regulation by season ID and regulation name
      operationId: getIdRegulations
      parameters:
        - $ref: "#/components/parameters/SeasonId"
        - $ref: "#/components/parameters/RegulationName"
      responses:
        "200":
          description: Regulation ID found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegulationIdResponse"
        "400":
          description: Invalid season_id or regulation_name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Regulation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RegulationId:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: "67d0ec1a412c1edc00d2117b"
    SeasonId:
      name: season_id
      in: path
      required: true
      schema:
        type: string
      example: "67d0ec1a412c1edc00d2117c"
    RegulationName:
      name: regulation_name
      in: path
      required: true
      schema:
        type: string
        enum:
          - "Age Regulation"
          - "Match Rules"
          - "Goal Rules"
          - "Ranking Rules"
      example: "Age Regulation"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Invalid email address"
      required:
        - status
        - message
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Operation successful"
        data:
          type: object
          nullable: true
          example: null
      required:
        - status
        - message
...
    Regulation:
      type: object
      properties:
        _id:
          type: string
          example: "67d0ec1a412c1edc00d2117b"
        season_id:
          type: string
          example: "67d0ec1a412c1edc00d2117c"
        regulation_name:
          type: string
          example: "Age Regulation"
        rules:
          type: object
          example:
            minAge: 18
            maxAge: 40
            minPlayersPerTeam: 11
            maxPlayersPerTeam: 20
            maxForeignPlayers: 5
        createdAt:
          type: string
          format: date-time
          example: "2025-04-08T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-04-08T12:00:00Z"
      required:
        - _id
        - season_id
        - regulation_name
        - rules
        - createdAt
        - updatedAt
    RegulationsResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Fetched regulations successfully"
        data:
          type: array
          items:
            $ref: "#/components/schemas/Regulation"
      required:
        - status
        - message
        - data
    CreateRegulationRequest:
      type: object
      properties:
        season_id:
          type: string
          example: "67d0ec1a412c1edc00d2117c"
        regulation_name:
          type: string
          enum:
            - "Age Regulation"
            - "Match Rules"
            - "Goal Rules"
            - "Ranking Rules"
          example: "Age Regulation"
        rules:
          type: object
          example:
            minAge: 18
            maxAge: 40
            minPlayersPerTeam: 11
            maxPlayersPerTeam: 20
            maxForeignPlayers: 5
      required:
        - season_id
        - regulation_name
        - rules
    RegulationResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Regulation found successfully"
        data:
          $ref: "#/components/schemas/Regulation"
      required:
        - status
        - message
        - data
    UpdateRegulationRequest:
      type: object
      properties:
        rules:
          type: object
          example:
            minAge: 18
            maxAge: 40
            minPlayersPerTeam: 11
            maxPlayersPerTeam: 20
            maxForeignPlayers: 5
      required:
        - rules
    RegulationIdResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Regulation id found"
        data:
          type: string
          example: "67d0ec1a412c1edc00d2117b"
      required:
        - status
        - message
        - data
...